hiding change plan assets on the frontend
