+++ rackcity/api/serializers/asset_serializers.py
@@ -16,6 +16,8 @@ from .change_plan_serializers import GetChangePlanSerializer
 from rackcity.api.serializers.fields import RCIntegerField
 import copy
 
+from ...utils.asset_changes_utils import get_changes_on_asset
+
 
 class AssetCPSerializer(serializers.ModelSerializer):
     """
@@ -305,6 +307,7 @@ class RecursiveAssetCPSerializer(serializers.ModelSerializer):
     related_asset = AssetSerializer()
     blades = serializers.SerializerMethodField()
     datacenter = serializers.SerializerMethodField()
+    mark_as_cp = serializers.SerializerMethodField()
 
     class Meta:
         model = AssetCP
@@ -337,6 +340,7 @@ class RecursiveAssetCPSerializer(serializers.ModelSerializer):
             "display_color",
             "memory_gb",
             "blades",
+            "mark_as_cp",
         )
 
     def get_mac_addresses(self, assetCP):
@@ -356,6 +360,10 @@ class RecursiveAssetCPSerializer(serializers.ModelSerializer):
 
     def get_datacenter(self, assetCP):
         return get_datacenter_of_asset(assetCP)
+    def get_mark_as_cp(self,assetCP):
+        if assetCP.related_asset:
+            return len(get_changes_on_asset(assetCP.related_asset,assetCP)) > 0 or assetCP.is_decommissioned
+        return True
 
 
 class GetDecommissionedAssetCPSerializer(serializers.ModelSerializer):
