+++ rackcity/utils/asset_utils.py
@@ -60,9 +60,6 @@ def get_existing_power_port(port_name, asset_id, change_plan=None):
 def does_asset_exist(asset_id, change_plan):
     return (
         AssetCP.objects.filter(id=asset_id, change_plan=change_plan.id).exists()
-        or AssetCP.objects.filter(
-            related_asset=asset_id, change_plan=change_plan.id
-        ).exists()
         or Asset.objects.filter(id=asset_id).exists()
     )
